ai-screening-bot/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pypdf==5.1.0
scikit-learn==1.5.2
openai>=1.40.0
pdfminer.six==20231228
